+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### [Unreleased]
 * Get Controller's route template from constant value
+* Added HTTP OPTIONS method
+* Remove obsolete status endpoint
 
 ### [0.1.0] - 2017-07-14
 Features in chronological order.